# Backend/swagger.yml
openapi: 3.0.3
info:
  title: Concert Manager API
  description: |
    API REST pour la gestion de concerts, groupes musicaux et salles.
    
    Cette API permet de :
    - Gérer les groupes musicaux (CRUD, import Excel)
    - Gérer les salles de concert (CRUD, import Excel)
    - Programmer et gérer les concerts (CRUD)
    - Consulter les statistiques du dashboard
    - Consulter les activités récentes
    

    **Format** : JSON

  version: 1.0.0
  contact:
    name: API Support
    email: support@concertmanager.com
  license:
    name: MIT
servers:
  - url: http://localhost:8005/api
    description: Serveur de développement local
  - url: https://api.concertmanager.com/api
    description: Serveur de production
...
paths:
...
  /salles:
    get:
      tags:
        - Salles
      summary: Liste des salles avec filtres et pagination
      description: |
        Récupère la liste des salles de concert avec possibilité de filtrer par :
        - city : Ville
        - minCapacity : Capacité minimale
        
        La pagination est gérée via les paramètres `page` et `limit`.
      operationId: listSalles
      parameters:
        - name: city
          in: query
          description: Filtrer par ville
          required: false
          schema:
            type: string
        - name: minCapacity
          in: query
          description: Capacité minimale
          required: false
          schema:
            type: integer
            minimum: 1
        - name: page
          in: query
          description: "Numéro de page (défaut: 1)"
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: "Nombre d'éléments par page (défaut: 10, max: 100)"
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Liste des salles récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalleListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Salles
      summary: Créer une nouvelle salle
      description: Crée une nouvelle salle de concert
      operationId: createSalle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalleCreateRequest'
            example:
              name: "Olympia"
              city: "Paris"
              capacity: 2000
              address: "28 Boulevard des Capucines, 75009 Paris"
      responses:
        '201':
          description: Salle créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /salles/{id}:
    get:
      tags:
        - Salles
      summary: Détails d'une salle
      description: Récupère les détails complets d'une salle par son ID
      operationId: getSalle
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la salle
          schema:
            type: integer
      responses:
        '200':
          description: Détails de la salle récupérés avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Salles
      summary: Modifier une salle
      description: Met à jour les informations d'une salle existante
      operationId: updateSalle
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la salle
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalleUpdateRequest'
      responses:
        '200':
          description: Salle modifiée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Salles
      summary: Supprimer une salle
      description: Supprime une salle et tous ses concerts associés
      operationId: deleteSalle
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la salle
          schema:
            type: integer
      responses:
        '200':
          description: Salle supprimée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /salles/import:
    post:
      tags:
        - Salles
      summary: Importer des salles depuis un fichier Excel
      description: |
        Importe plusieurs salles depuis un fichier Excel (.xlsx, .xls).
        
        Format attendu du fichier Excel :
        - Colonne A : Nom
        - Colonne B : Ville
        - Colonne C : Capacité
        - Colonne D : Adresse
      operationId: importSalles
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Fichier Excel (.xlsx ou .xls)
      responses:
        '200':
          description: Import réussi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "5 salles importées avec succès"
                  data:
                    type: object
                    properties:
                      imported:
                        type: integer
                        example: 5
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            row:
                              type: integer
                            message:
                              type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Salle:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de la salle
          example: 1
        name:
          type: string
          description: Nom de la salle
          example: "Olympia"
        city:
          type: string
          description: Ville
          example: "Paris"
        capacity:
          type: integer
          description: Capacité (nombre de places)
          minimum: 1
          example: 2000
        address:
          type: string
          description: Adresse complète
          example: "28 Boulevard des Capucines, 75009 Paris"
    SalleCreateRequest:
      type: object
      required:
        - name
        - city
        - capacity
        - address
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Olympia"
        city:
          type: string
          minLength: 1
          maxLength: 255
          example: "Paris"
        capacity:
          type: integer
          minimum: 1
          example: 2000
        address:
          type: string
          minLength: 1
          example: "28 Boulevard des Capucines, 75009 Paris"
    SalleUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/SalleCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              description: ID de la salle à modifier
    SalleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Salle'
    SalleListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Salle'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ===========================================
    # CONCERT SCHEMAS
    # ===========================================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Page actuelle
          example: 1
        limit:
          type: integer
          description: Nombre d'éléments par page
          example: 10
        total:
          type: integer
          description: Nombre total d'éléments
          example: 100
        totalPages:
          type: integer
          description: Nombre total de pages
          example: 10
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Opération réussie"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Message d'erreur détaillé"
  responses:
    BadRequest:
      description: Requête invalide (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Les données fournies sont invalides"
    NotFound:
      description: Ressource introuvable (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Ressource introuvable"
    InternalServerError:
      description: Erreur serveur (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Une erreur interne s'est produite"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentification par token JWT.
        Actuellement en mode mock pour le développement.
        Format: `Bearer {token}`
security:
  - BearerAuth: []